# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My-Trip API Documentation",
        "description": "API documentation for My-Trip Platform",
        "contact": {
            "email": "support@mytrip.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/wjhtak-site/public",
            "description": "Development Server"
        },
        {
            "url": "https://blueviolet-hummingbird-437500.hostingersite.com/public",
            "description": "Production Server 1"
        },
        {
            "url": "https://lavenderblush-bear-243464.hostingersite.com/public/",
            "description": "Production Server 2"
        }
    ],
    "paths": {
...
        "/api/v1/countries": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "summary": "Get all countries",
                "description": "Retrieve a list of all active countries with their flags and codes.",
                "operationId": "getCountries",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Countries retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Countries retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Saudi Arabia"
                                                    },
                                                    "name_ar": {
                                                        "type": "string",
                                                        "example": "المملكة العربية السعودية"
                                                    },
                                                    "name_en": {
                                                        "type": "string",
                                                        "example": "Saudi Arabia"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "SA"
                                                    },
                                                    "phone_code": {
                                                        "type": "string",
                                                        "example": "966"
                                                    },
                                                    "flag": {
                                                        "type": "string",
                                                        "example": "http://example.com/flags/sa.png"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cities": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "summary": "Get all cities",
                "description": "Retrieve a list of all active cities, optionally filtered by country.",
                "operationId": "getCities",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    },
                    {
                        "name": "country_id",
                        "in": "query",
                        "description": "Filter cities by country ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cities retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cities retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Riyadh"
                                                    },
                                                    "name_ar": {
                                                        "type": "string",
                                                        "example": "الرياض"
                                                    },
                                                    "name_en": {
                                                        "type": "string",
                                                        "example": "Riyadh"
                                                    },
                                                    "country": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Saudi Arabia"
                                                            },
                                                            "flag": {
                                                                "type": "string",
                                                                "example": "http://example.com/flags/sa.png"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/banners": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "summary": "Get banners",
                "description": "Retrieve a list of promotional banners for the home screen.",
                "operationId": "getBanners",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banners retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Banners retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Summer Sale"
                                                    },
                                                    "title_ar": {
                                                        "type": "string",
                                                        "example": "عروض الصيف"
                                                    },
                                                    "title_en": {
                                                        "type": "string",
                                                        "example": "Summer Sale"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Get 50% off on flights"
                                                    },
                                                    "description_ar": {
                                                        "type": "string",
                                                        "example": "احصل على خصم 50% على الرحلات"
                                                    },
                                                    "description_en": {
                                                        "type": "string",
                                                        "example": "Get 50% off on flights"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "http://example.com/banners/1.jpg"
                                                    },
                                                    "link": {
                                                        "type": "string",
                                                        "example": "https://example.com/promo"
                                                    },
                                                    "order": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/faqs": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "summary": "Get all FAQs",
                "description": "Retrieve a list of all active frequently asked questions.",
                "operationId": "getFaqs",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "FAQs retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "question": {
                                                        "type": "string",
                                                        "example": "How to book?"
                                                    },
                                                    "answer": {
                                                        "type": "string",
                                                        "example": "You can book via..."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token to access protected routes",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}